llama_test1_requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
bibtexparser==2.0.0b8
bitsandbytes==0.45.0
blinker==1.9.0
braceexpand==0.1.7
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ChatTTS==0.0.5
click==8.1.8
clldutils==3.24.0
cmake==3.31.2
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
csvw==3.5.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
dlinfo==1.2.1
editdistance==0.8.1
einops==0.8.0
einx==0.3.0
encodec==0.1.1
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
frozendict==2.4.6
fsspec==2024.12.0
funasr==1.1.6
gradio==3.47.1
gradio_client==0.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_resources==6.5.2
inflect==7.5.0
iopath==0.1.10
isodate==0.7.2
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldiio==2.18.0
kiwisolver==1.4.8
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
lit==18.1.8
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.0
more-itertools==10.5.0
moviepy==1.0.2
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.21.1
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.7.0.72
orjson==3.10.14
oss2==2.19.1
packaging==24.2
pandas==2.2.3
peft==0.11.0
phonemizer==3.3.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
proglog==0.1.10
progressbar2==4.5.0
protobuf==5.29.3
psutil==5.9.4
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pylatexenc==2.10
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-utils==3.9.1
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0
rdflib==7.1.1
referencing==0.35.1
regex==2022.10.31
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
segments==2.2.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.3
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.6.13
tokenizers==0.13.2
torch==2.1.0
torch-complex==0.4.4
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.64.1
transformers==4.30.0
triton==2.1.0
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vector-quantize-pytorch==1.21.1
visual-genome==1.1.1
vocos==0.1.0
webdataset==0.2.48
websockets==11.0.3
Werkzeug==3.1.3
unidecode
flask